+++ CHANGELOG.md
@@ -4,12 +4,17 @@
 
 ### (Work in Progress) Version 4.4.0
 
+#### Enterprise Policies
+
+- new policy to enable or disable captive portal support (Firefox 67+), fixes
+  [#69](https://github.com/cadeyrn/enterprise-policy-generator/issues/69)
+
+#### Dependencies
+
 - updated eslint from version 5.12.1 to 5.15.2
 - updated eslint-plugin-compat from version 2.6.3 to 3.0.1
 - updated stylelint-order from version 2.0.0 to 2.1.0
 - updated web-ext from version 2.9.3 to 3.0.0
-
-#### Dependencies
 
 [All Changes](https://github.com/cadeyrn/enterprise-policy-generator/compare/v4.3.0...master)<br />